requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.20.0
asyncpg==0.30.0
pydantic==2.11.7
python-dotenv==1.1.1
PyJWT==2.10.1
slowapi==0.1.9
psutil==6.1.1
